fix:useTaskでの仮データ削除/fix:インデントの整理
